Вносит исправления
