+++ js/get-ordered-pictures.js
@@ -1,0 +1,23 @@
+const indexList = [];
+
+const getOrderedPictures = (startNumber, path, maxCount) => {
+  const indexElement = indexList[indexList.length - 1] + 1;
+
+  for (let i = startNumber; i <= maxCount; i++) {
+
+    if (indexList.length < 1) {
+      indexList.push(startNumber);
+      return `${path}0${startNumber}.png`;
+
+    } else if (indexList.length >= 9) {
+      indexList.push(indexElement);
+      return `${path}${indexElement}.png`;
+
+    } else {
+      indexList.push(indexElement);
+      return `${path}0${indexList[indexList.length - 1]}.png`;
+    }
+  }
+};
+
+export {getOrderedPictures};
