+++ js/main.js
@@ -1,5 +1,3 @@
-import './get-ordered-entity.js';
+import './get-ordered-pictures.js';
 import './get-random-entity.js';
-import './data.js';
 import './get-similar-entity.js';
-import './remove-block.js';
